Fix ys in Pollard's rho algorithm
Previously ys simply tracked the previous value of y. This doesn't make
a lot of sense because it means the final loop, meant to protect against
finding multiple factors at once, will only run for one iteration before
encountering the known non-1 value for G. In Brent's paper it's set
before the size-m loop instead of in its body. This is the more logical
location because the final loop is a backtracking step to go over the
last m or so values with the more accurage gcd instead of the faster
multiplication.