Allow DataStructures v0.18
DataStructures.jl is a core package and v0.18 was released in Aug 17, 2020. This means that pretty much any package made after that won't be compatible with Primes.jl in its latest form, so this shows up as a major ecosystem update blocker. We should really make sure that this kind of repo has CompatHelper.jl on it.